バグ修正

+++ MarathonRunner.Core/Runners/Callbacks/ScoreLogger.cs
@@ -42,8 +42,8 @@ public class ScoreLogger : IRunnerCallback
     protected virtual void LogSingleResult(int seed, TestCaseResult result)
     {
         var elapsed = (int)result.Elapsed.TotalMilliseconds;
-        var rate = result.Score * 100 / _referenceScore;
-        var average = _totalScore * 100 / (_referenceScore * _completedCaseCount);
+        var rate = (double)result.Score * 100 / _referenceScore;
+        var average = (double)_totalScore * 100 / (_referenceScore * _completedCaseCount);
 
         _logger.LogInformation("case {seed:000000} | completed: {count,5} / {cases,5} | {elapsed,5:#,##0} ms | {score,11:#,##0} | rate: {rate,7:0.000}% | average: {average,7:0.000}%",
             seed, _completedCaseCount, _plannedCaseCount, elapsed, result.Score, rate, average);
